requirements.txt
autopep8==2.0.4
attrs==23.1.0
backcall==0.2.0
certifi==2023.11.17
chardet==5.2.0
decorator==5.1.1
Django==4.2.7
django-environ==0.11.2
django-extensions==3.2.3
django-redis==5.4.0
flake8==6.1.0
flake8-commas==2.1.0
flake8-mypy==17.8.0
gevent==23.9.1
greenlet==3.0.1
gunicorn==21.2.0
idna==3.6
ipython==8.18.1
ipython-genutils==0.2.0
jedi==0.19.1
mccabe==0.7.0
mypy==1.7.1
pexpect==4.9.0
pickleshare==0.7.5
Pillow==10.1.0
prompt-toolkit==3.0.41
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycodestyle==2.11.1
pyflakes==3.1.0
Pygments==2.17.2
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
rope==1.11.0
simplegeneric==0.8.1
six==1.16.0
traitlets==5.14.0
typed-ast==1.5.5
urllib3==2.1.0
wcwidth==0.2.12
Werkzeug==3.0.1
djangorestframework==3.14.0
django-htmx
django_extensions
django-filter~=23.5
python-dotenv~=1.0.0